archive/Pi Code/requirements.txt
pyserial==3.5
RPi.GPIO==0.7.1
pybluez==0.23
bleak==0.21.1